requirements.txt
# numpy==1.19.2
pandas==1.1.3
PyYAML==5.3.1
matplotlib==3.3.2
scikit-learn==0.23.2
transformers==3.5.0
seaborn==0.11.0
torch==1.7.0
tqdm==4.51.0
numpy==1.20
dcor
hdbscan
networkx
# pandas
spacy

nolds
scipy
nltk
# transformers
sklearn
statsmodels
spacy_syllables
dtw-python
